backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
eyed3
python-dotenv
